refactor: CI/CD를 위해 통과하는 (임시로) 테스트 코드로 리팩토링
추후 주기적으로 리팩토링하며 코드 교체

+++ src/main/java/com/roki/fashionfeed/web/FeedShareApiController.java
@@ -1,7 +1,9 @@
 package com.roki.fashionfeed.web;
 
+import com.roki.fashionfeed.config.auth.LoginUser;
+import com.roki.fashionfeed.config.auth.dto.SessionUser;
 import com.roki.fashionfeed.service.ShareService;
-import com.roki.fashionfeed.web.dto.LikefullSaveRequestDto;
+import com.roki.fashionfeed.service.UserService;
 import com.roki.fashionfeed.web.dto.ShareSaveRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -9,20 +11,16 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-
 @RequiredArgsConstructor
 @RestController
 public class FeedShareApiController {
     private final ShareService shareService;
+    private final UserService userService;
 
-    @PutMapping("/api/{feedId}/shares")
-    public Long save(@PathVariable Long feedId, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        Object objUserId = session.getAttribute("sessionUser");
-        if (objUserId != null) {
-            Long userId = (long) (int) objUserId;
+    @PutMapping("/api/{feedId}/shares") //todo: 마찬가지로 feeds/{feedsId} 되야할 듯!
+    public Long save(@PathVariable Long feedId, @LoginUser SessionUser user) {
+        if (user != null) {
+            Long userId = userService.findUserIdByEmail(user.getEmail());
             return shareService.save(feedId, new ShareSaveRequestDto(userId));
         }
         // 유저 아이디가 없는 경우 발생
@@ -30,10 +28,13 @@ public class FeedShareApiController {
     }
 
     @DeleteMapping("/api/{feedId}/shares")
-    public Long delete(@PathVariable Long feedId, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-        Long userId = (long) (int) session.getAttribute("sessionUser");
-        shareService.delete(feedId, userId);
-        return userId;
+    public Long delete(@PathVariable Long feedId, @LoginUser SessionUser user) {
+        if (user != null) {
+            Long userId = userService.findUserIdByEmail(user.getEmail());
+            shareService.delete(feedId, userId);
+            return userId;
+        }
+
+        return -1L;
     }
 }
